Room Mine update
It has been implemented the usage of mine key. In Mine Room, it has been simulated a locked door. To open it, is required the mine key to unlock earth altar.
Moreover, has been implemented story to guide the player that needs to find a key in order to continue.

In Room class, setNeighbor has been implemented in order to change the neighbor in the middle of the game.

In World class, has been added the description of all rooms.